Use getFinalPrice method for compatibility with all product types [Update #2]
Use the $item object in order to retrieve the price using the getPrice() method as the quote item is storing the right price value which should ensure compatibility with all product types. The part comparing with the special price isn't required with this approach as the price registered within the quote item is the final price. This price is also right for configurable products having their price modified through options.